added UI html for approve volunteers, created api methods, edited volunteerdao to implement CU methods to move insert volunteer usernames into users table, created approvedVolunteerDTO to handle the simple logic. Encountered Bug where the approve method on UI side is not working right and crashes the server. Need help

+++ java/src/main/java/com/techelevator/model/Volunteer.java
@@ -6,7 +6,7 @@ public class Volunteer {
     private int id;
     private String name;
     private String email;
-    private int phone;
+    private long phone;
     private int applicationStatusId = 1;
 
     public String getUsername() {
@@ -54,7 +54,7 @@ public class Volunteer {
         this.email = email;
     }
 
-    public int getPhone() {
+    public long getPhone() {
         return phone;
     }
